// src/model/marvelResponse.ts
// Generated by https://quicktype.io

export interface MarvelResponseData<T> {
    code:            number;
    status:          string;
    copyright:       string;
    attributionText: string;
    attributionHTML: string;
    etag:            string;
    data:            MarvelData<T>;
}

export interface MarvelData<T> {
    offset:  number;
    limit:   number;
    total:   number;
    count:   number;
    results: T[];
}